refactor(Leagues): cleaned up the main page view
