+++ package.json
@@ -27,7 +27,6 @@
 		"postcss-load-config": "^4.0.1",
 		"prettier": "^2.8.0",
 		"prettier-plugin-svelte": "^2.8.1",
-		"puppeteer": "^19.8.0",
 		"svelte": "^3.54.0",
 		"svelte-preprocess": "^4.10.7",
 		"tailwindcss": "^3.1.5",
