--- pupeteer.config.cjs
@@ -1,9 +1,0 @@
-const { join } = require('path');
-
-/**
- * @type {import("puppeteer").Configuration}
- */
-module.exports = {
-	// Changes the cache location for Puppeteer.
-	cacheDirectory: join(__dirname, '.cache', 'puppeteer')
-};
