+++ src/lib/league/components.js
@@ -1,0 +1,4 @@
+import Fixtures from './Fixtures.svelte';
+import Table from './Table.svelte';
+
+export { Table, Fixtures };
